wombat/driver-tests/requirements.txt
rpi-lgpio
adafruit-circuitpython-pca9685
pyserial